Added stack build files. Added .gitignore for stack build stuff
Modified OpenRecords.hs to have instance of IsLabel so the records can be used with the GHC
8.x overloaded labels.

Also exposed a few of the "hidden" constructors as I need them in matching type constraints in some
class/type family instances. This is down mostly to the fact that GHC is not reducing them as one
might hope!